+++ src/main/java/net/minecraft/client/renderer/entity/RendererLivingEntity.java
@@ -107,7 +107,7 @@ public abstract class RendererLivingEntity extends Render
             float var11 = this.interpolateRotation(par1EntityLivingBase.prevRotationYawHead, par1EntityLivingBase.rotationYawHead, par9);
             float var13;
 
-            if (Nebula.getInstance().getRotations().isRotating() && renderManager.livingPlayer.equals(Minecraft.getMinecraft().thePlayer)) {
+            if (Nebula.getInstance().getRotations().isRotating() && par1EntityLivingBase.equals(Minecraft.getMinecraft().thePlayer)) {
                 var11 = this.interpolateRotation(par1EntityLivingBase.prevClientYaw, par1EntityLivingBase.clientYaw, par9);
             }
 
@@ -137,7 +137,7 @@ public abstract class RendererLivingEntity extends Render
 
             float var26 = par1EntityLivingBase.prevRotationPitch + (par1EntityLivingBase.rotationPitch - par1EntityLivingBase.prevRotationPitch) * par9;
 
-            if (Nebula.getInstance().getRotations().isRotating() && renderManager.livingPlayer.equals(Minecraft.getMinecraft().thePlayer)) {
+            if (Nebula.getInstance().getRotations().isRotating() && par1EntityLivingBase.equals(Minecraft.getMinecraft().thePlayer)) {
                 var26 = par1EntityLivingBase.prevClientPitch + (par1EntityLivingBase.clientPitch - par1EntityLivingBase.prevClientPitch) * par9;
             }