+++ CHANGELOG.md
@@ -37,6 +37,8 @@
 + Added "Reset VL" module
 + Added "Nuker" module
 + Fixed Flatten to sneak to place on blocks that need to be sneaken on
++ Fixed other entities render rotations from being 0'd out when using spoof rotations
++ 
 ```
 
 ---